tests(vitest): moved vitest files into separete folder
