+++ vite.config.ts
@@ -25,5 +25,12 @@ export default defineConfig({
 		environment: "jsdom",
 		globals: true,
 		setupFiles: "./src/tests/setup.ts",
+		coverage: {
+			provider: "v8",
+			reporter: ["text", "json", "html"],
+			include: ["src/**/*.{ts,tsx}"],
+			exclude: ["src/tests/**/*", "node_modules/**/*"],
+		},
+		include: ["./src/tests/vitest/**/*.{test,spec}.{ts,tsx}"],
 	},
 });
